Add implicit Nil returns if empty return
This is quite a scuffed way to do it. The way it tries to see if there
is an expression after the return is not great and can probably miss
some cases. But it will work for most normal cases, and some cases you
might not expect, but probably wont try :D As it is an expression and
not a statements it can be in many strange places...