Add first integration test
It is not nice that we have to run a subprocess command. Ideally we
would just call one of the functions from lib.rs, and redirect the
output from stdout to the test for inspection. The way to do this seems
to be to pass some writer to all places and use it instead of println,
but that is a lot of work, and clutters the code, so I leave it for now.