Add tuple lvalues to match against iterators
A tuple as an lvalue will match each of its contained lvalues to one of
the rvalues in the rvalue iterator. If any of the two sides are of
different lengths it errors.

Also solve aoc day 4 with it! Very brute force with sets, but went
quite fast! Makes me feel good about the languges chances.