Add semicolon elision
Quite simple, allow elision for `if`, `while`, `for`, and function decl
if they end with a block. Also always accept match and blocks.

Tricky part was to seperate indexing and calls over lines. Now the
language is whitespace aware in that calls and indexing must follow
immediately after the thing being called/indxed. So `{x}[1];` is the same
as `x[1]`, while `{x} [1];` is two statements in sequence `{x}` and
`[1]`. This is a bit of a hard pill to swallow, but makes sense in how
you write code.