Solve day 18
It is quite a simple day, but for part two I wanted to do a dfs and
instead ran out of stack frames around 700 function calls in. I guess
that for every function call we have severeal ones in Rust, but it
is still very bad. This could be way better in a vm solution, so I
saved that approach so we can test if out vm can do it.