Solve all of aoc 2022!
This proves the language is usable for its purpose. The stdlib lacks
a few things, but that can always be the case. Furthermore, I think
it would be nice to have exceptions (both yourself? and for num ops),
as well as a decorator for caching (arbitrary ones?). I have added
TODO notes where more things felt needed.