Add procedural macro for enum <-> u8
It seems to work, but the main issue is that my file organization is
completely chaos. Another crate should probably not be in the src dir.
Looking at other crates, maybe I should break out all parts of the repo
to different crates. Such as _scanner, parses, ast_interpreter, vm_.