Implement closures
Every time a a function is defined, it is now a
closure which wrapps a function. This closure
keeps track of contained upvalues. Furthermore,
values which will be captured are stored as
pointers on the stack to a value on the heap,
with special instructions for working with them.

This seems to work for a toy example, but it does
not have real tests, so probably contains a lot of
bugs.