Add match expressions!
It would be good with more tests, for example on dicts, strings, and
testing that the output is valid. However, this was the last 'real'
feature to be added to the language before its first release! Now we
just want a dict and more built in functions (maybe an way to import
other files as well).